moved spec for quick submit from the landing page suite to its own suite
