+++ site/gatsby-site/playwright/e2e/landingPage.spec.ts
@@ -1,8 +1,5 @@
 import { expect } from '@playwright/test';
-import { conditionalIntercept, waitForRequest, test } from '../utils';
-import { format } from 'date-fns';
-
-const now = new Date();
+import { test } from '../utils';
 
 test.describe('The Landing page', () => {
   test('successfully loads', async ({ page }) => {
@@ -29,28 +26,6 @@ test.describe('The Landing page', () => {
     await page.locator('[data-cy="Waking Up Foundation-image"]').click();
     await expect(page.locator('[data-cy="sponsor-modal"] h5', { hasText: 'Waking Up Foundation' })).toBeVisible();
   });
-
-  test('Should submit a report through the Quick Add form', async ({ page }) => {
-    await conditionalIntercept(
-      page,
-      '**/graphql',
-      (req) => req.postDataJSON().operationName == 'InsertQuickAdd',
-      { data: { insertOneQuickadd: { __typename: 'Quickadd', _id: '6271a62068d0c59a372b0c09' } } },
-      'InsertQuickAdd'
-    );
-
-    await page.goto('/apps/submit');
-    await page.locator('[data-cy="quick-add"] [name="quickAddUrl"]').fill('https://example.com');
-    await page.locator('[data-cy="quick-add"] [type="submit"]').click();
-
-    const insertQuickAddRequest = await waitForRequest('InsertQuickAdd');
-    const variables = insertQuickAddRequest.postDataJSON().variables;
-    expect(variables.data.url).toBe('https://example.com/');
-    expect(variables.data.date_submitted).toBe(format(now, 'yyyy-MM-dd'));
-
-    await expect(page.locator('.tw-toast')).toContainText('Report successfully added to review queue. You can see your submission here.');
-  });
-
 
   test.skip('Should redirect to the account page when logged in',
     async ({ page, skipOnEmptyEnvironment, login }) => {
